Moved Base64 and BiFunction from java util package to dari util for Java7 compliance.
