+++ util/src/main/java/com/psddev/dari/util/Base64.java
@@ -23,7 +23,7 @@
  *
  */
 
-package java.util;
+package com.psddev.dari.util;
 
 import java.io.FilterOutputStream;
 import java.io.InputStream;
@@ -31,6 +31,8 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
+import java.util.Objects;
 
 /**
  * This class consists exclusively of static methods for obtaining
